requirements.txt
numpy
rich
einops
matplotlib
glob2
pyyaml
scipy
tqdm
opencv-python
pandas